end-to-end_project/services/text_similarity.py
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity

def text_similarity_score(text1, text2):
    vectorizer = TfidfVectorizer(stop_words="english")

    tfidf = vectorizer.fit_transform([text1, text2])
    similarity = cosine_similarity(tfidf[0:1], tfidf[1:2])

    return similarity[0][0]
