end-to-end_project/requirements.txt
flask
flask_sqlalchemy
Flask-Bcrypt
flask_jwt_extended
flask_cors
PyMySQL
cryptography
python-dotenv
Flask-Migrate
Flask-APScheduler
scikit-learn